requirements.txt
prometheus_client
numpy
pandas
cassandra-driver
requests
